Rework and split to the components
