--- src/components/ChatComponents/UserList.tsx
@@ -1,35 +1,0 @@
-// import React from 'react';
-// import {
-//     List,
-//     ListItem,
-//     ListItemAvatar,
-//     Avatar,
-//     ListItemText,
-//     Typography
-// } from '@mui/material';
-//
-// const UserList = () => {
-//     return (
-//         <List>
-//             <ListItem alignItems="flex-start">
-//         <ListItemAvatar>
-//             <Avatar alt="Remy Sharp" src="/static/images/avatar/1.jpg" />
-//         </ListItemAvatar>
-//         <ListItemText
-//     primary="Brunch this weekend?"
-//     secondary={
-//     <React.Fragment>
-//     <Typography sx={{ display: 'inline' }} component="span" variant="body2" color="text.primary">
-//         Ali Connors
-//     </Typography>
-//     {" — I'll be in your neighborhood doing errands this…"}
-//     </React.Fragment>
-// }
-//     />
-//     </ListItem>
-//     {/* Additional user cells can be added here */}
-//     </List>
-// );
-// };
-//
-// export default UserList;
