+++ src/utils/types.ts
@@ -1,0 +1,5 @@
+export interface Message {
+  id: number;
+  text: string;
+  sender: string;
+}
